bash: ConEmuC: command not found
